correciones

+++ clases/Estados.cs
@@ -8,7 +8,7 @@ namespace clases
 {
     public class Estados
     {
-        public Int16      Id      { get; set; }
+        public int Id      { get; set; }
         public string   Estado  { get; set; }
     }
 }
